+++ demo/index.html
@@ -5,7 +5,7 @@
 		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
 		<script type="module">
 			import '@brightspace-ui/core/components/demo/demo-page.js';
-			import '../awards.js';
+			import '../src/components/awards.js';
 		</script>
 		<title>d2l-awards</title>
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
